CNN_image_classifier/requirements-docker.txt
fastapi==0.128.0
uvicorn==0.40.0
tensorflow-cpu==2.20.0
keras==3.13.0
pillow==12.1.0
numpy==2.4.0
pydantic==2.12.5
python-multipart==0.0.21
